pyproject.toml
[project]
name = "kc-cert-tracker"
version = "0.1.0"
description = "Track IT certification demand in Kansas City job postings"
requires-python = ">=3.11"
dependencies = [
    "requests",
    "psycopg2-binary",
    "pandas",
    "python-dotenv",
    "fastapi",
    "uvicorn",
    "boto3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]